+++ packages/core/src/stream-loading/anchor-timestamp-extractor.ts
@@ -1,14 +1,14 @@
-import {
-  AnchorValidator,
+import type {
   AppliableStreamLog,
   CommitData,
-  CommitType,
   DiagnosticsLogger,
   UnappliableStreamLog,
 } from '@ceramicnetwork/common'
-import { CID } from 'multiformats/cid'
+import { CommitType } from '@ceramicnetwork/common'
+import type { CID } from 'multiformats/cid'
+import type { AnchorValidator } from '../anchor/anchor-service.js'
 
-interface IpfsRecordLoader {
+type IpfsRecordLoader = {
   retrieveFromIPFS(cid: CID | string, path?: string): Promise<any>
 }