+++ packages/stream-tile-handler/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.32.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile-handler@2.31.0...@ceramicnetwork/stream-tile-handler@2.32.0-rc.0) (2023-10-11)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
+# [2.31.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile-handler@2.31.0-rc.0...@ceramicnetwork/stream-tile-handler@2.31.0) (2023-10-11)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
 # [2.31.0-rc.0](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile-handler@2.30.0...@ceramicnetwork/stream-tile-handler@2.31.0-rc.0) (2023-09-20)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-tile-handler